PythonBackend/flask-face-recognition/requirements.txt
Flask
pymongo
face_recognition
opencv-python
python-dotenv
PyJWT
flask-cors
